Added Vat Category Maint

+++ RetailGInvoices/Data/VatRate.cs
@@ -11,6 +11,6 @@ namespace RetailGInvoices.Data
         [Key]
         public int Id { get; set; }
         public string Description { get; set; }
-        public float Rate { get; set; }
+        public Decimal Rate { get; set; }
     }
 }
